Refactoring mobile video upload

+++ src/mobile/TB.DanceDance.Mobile.Library/Services/Network/UploadProgressEvent.cs
@@ -3,6 +3,6 @@
 public record UploadProgressEvent()
 {
     public required string FileName { get; init; }
-    public required int SendBytes { get; init; }
+    public required long SendBytes { get; init; }
     public required long FileSize { get; init; }
 }
